Added base FlameAnimationController
