+++ packages/flame/lib/src/effects/new_move_effect.dart
@@ -1,13 +1,13 @@
 
 import 'package:vector_math/vector_math_64.dart';
-import 'flame_animation_controller.dart';
+import 'standard_animation_controller.dart';
 import 'transform2d_effect.dart';
 
 
 class NewMoveEffect extends Transform2DEffect {
   NewMoveEffect.to({
     required Vector2 destination,
-    required FlameAnimationController controller,
+    required StandardAnimationController controller,
   })  : _absolute = true,
         _offset = destination.clone(),
         _lastProgress = 0,
@@ -15,7 +15,7 @@ class NewMoveEffect extends Transform2DEffect {
 
   NewMoveEffect.by({
     required Vector2 offset,
-    required FlameAnimationController controller,
+    required StandardAnimationController controller,
   })  : _absolute = false,
         _offset = offset.clone(),
         _lastProgress = 0,
